requirements.txt
python-dotenv
openai
pandas
pydantic
griffe
openpyxl
